fix(rider-session): sửa crash và chặn cảnh báo quá giờ vô nghĩa
Rà soát lại phần cảnh báo quá giờ đêm và quá giờ xe số tự động sau khi chạy thử
trên thiết bị.

Crash khi kết thúc phiên: khối cập nhật ô giờ đêm chạy mỗi giây và gọi getString
nên cần context, mà kết thúc phiên thì fragment bị detach ngay trong lúc một
tick còn đang chờ trên Main, và clockLogicBlock không có
CoroutineExceptionHandler. Guard isAdded cho cả các khối cảnh báo 24h, giảng
viên và 4 giờ vốn đã có cùng rủi ro, trừ khối tự động đăng xuất để không bỏ qua
việc đăng xuất theo quy định.

Chặn ClassCastException khi cập nhật app: giờ khung đêm trước đây lưu shared-pref
bằng putInt, nay là putFloat để nhận giờ lẻ, nên getFloat trên khoá cũ sẽ ném
lỗi ngay trong khối chạy mỗi giây. Đọc bằng readHourPreference chịu được cả hai
kiểu.

Chỉ cảnh báo khi con số còn tăng được:
- giờ đêm chỉ cảnh báo khi hiện tại nằm trong khung đêm, thêm
  Utils.isInNightWindow. Trước đây học viên đã vượt ngưỡng mở phiên ban ngày sẽ
  bị cảnh báo 5 phút một lần suốt phiên trong khi con số đứng im
- giờ xe số tự động chỉ cảnh báo khi isAutomaticTransmission = true, không còn
  cảnh báo quá giờ xe số tự động trong lúc đang lái xe số sàn

Nhận khung giờ lẻ: nightFromHour và nightToHour đổi sang Double, bỏ toInt ở
mapper, nên khung kiểu 18.5 nghĩa là 18h30 tính đúng thay vì lệch 30 phút.

Ngưỡng nhấp nháy không còn âm khi Max nhỏ hơn 15 phút.

Đã kiểm chứng 16 ca cho phép giao khung đêm, gồm giờ lẻ và hai biên đầu cuối
khung. Cài và chạy trên FP-08 không crash.

+++ app/src/main/java/com/hc/dat/model/SessionTimeLimitInfo.kt
@@ -10,10 +10,10 @@ package com.hc.dat.model
 data class SessionTimeLimitInfo(
     /** Số giây đã học trong khung đêm trước phiên hiện tại, do server tổng hợp. */
     val nightTime: Double = 0.0,
-    /** Giờ bắt đầu khung đêm trong ngày, 0..23. */
-    val nightFromHour: Int = 0,
-    /** Giờ kết thúc khung đêm trong ngày, 0..23; bằng nightFromHour là khung rỗng. */
-    val nightToHour: Int = 0,
+    /** Giờ bắt đầu khung đêm trong ngày, nhận cả giờ lẻ như 18.5 = 18h30. */
+    val nightFromHour: Double = 0.0,
+    /** Giờ kết thúc khung đêm trong ngày; bằng nightFromHour là khung rỗng. */
+    val nightToHour: Double = 0.0,
     /** Ngưỡng giờ đêm tối đa theo giây, 0 nghĩa là server không giới hạn. */
     val nightTimeMax: Double = 0.0,
     /** Số giây đã học trên xe số tự động trước phiên hiện tại. */
